Move Head into foundation
