+++ src/lib/foundation/index.ts
@@ -1,6 +1,7 @@
 export { default as Button } from "./Button.svelte";
 export { default as Chooser } from "./Chooser.svelte";
 export { default as Column } from "./Column.svelte";
+export { default as Head } from "./Head.svelte";
 export { default as LoadingIndicator } from "./LoadingIndicator.svelte";
 export { default as Row } from "./Row.svelte";
 export { default as TextInput } from "./TextInput.svelte";
